Modify prince column type for float

+++ db/migrate/20220829192510_add_details_to_food.rb
@@ -2,7 +2,7 @@ class AddDetailsToFood < ActiveRecord::Migration[7.0]
   def change
     add_column :foods, :name, :string
     add_column :foods, :measurement_unit, :string
-    add_column :foods, :price, :integer
+    add_column :foods, :price, :float
     add_column :foods, :quantity, :integer
   end
 end
